Reamped install script
The install script is now more handholdy.

Upon running the installer it will check if the system is debian based.
If so it will install the required dependencies, if not it will inform the user of the dependency.

The installer will check if the compile and install of the rust program was successful.

The installer will only add the icon, if the icon is not already added to the cdg list.

Upon installer exit it will inform the user about how many errors occured during installation.