Added a quick refresh period after the door opens
Sometimes the door opens when someone just goes in to grab something.
To avoiid people rushing OV in this scenario because the program said it
opened, the rate of API checks are 4x in the minute after the door opens.

It hasn't been tested yet, but I believe it should work.